+++ src/sql/setup.sql
@@ -8,7 +8,7 @@ CREATE TABLE accounts
     name VARCHAR(20) NOT NULL,
     password VARCHAR(25) NOT NULL,
     role ENUM('USER','BOOKMAKER', 'ADMIN') NOT NULL,
-    balance float DEFAULT 0.0 NOT NULL,
+    balance DECIMAL(12,4) DEFAULT 0.0 NOT NULL,
     is_blocked BOOLEAN DEFAULT 0 NOT NULL,
     PRIMARY KEY(id)
 );
@@ -32,8 +32,8 @@ CREATE TABLE matches
     tournament VARCHAR(15) NOT NULL,
     commission FLOAT DEFAULT 0.0 NOT NULL,
     is_closed BOOLEAN DEFAULT 0 NOT NULL,
-    first_team_bets FLOAT DEFAULT 0.0 NOT NULL,
-    second_team_bets FLOAT DEFAULT 0.0 NOT NULL,
+    first_team_bets DECIMAL(12,4) DEFAULT 0.0 NOT NULL,
+    second_team_bets DECIMAL(12,4) DEFAULT 0.0 NOT NULL,
     FOREIGN KEY (first_team_id)  REFERENCES teams (id) ON DELETE CASCADE,
 	FOREIGN KEY (second_team_id) REFERENCES teams (id) ON DELETE CASCADE,
     PRIMARY KEY(id)
@@ -45,8 +45,8 @@ CREATE TABLE bets
 	match_id   BIGINT       NOT NULL,
     account_id BIGINT       NOT NULL,
     team_id    BIGINT       NOT NULL,
-    money_bet float NOT NULL,
-    money_received float DEFAULT 0.0 NOT NULL,
+    money_bet      DECIMAL(12,4) NOT NULL,
+    money_received DECIMAL(12,4) DEFAULT 0.0 NOT NULL,
 	bet_date DATETIME NOT NULL,
 	PRIMARY KEY(id),
 	FOREIGN KEY (match_id) REFERENCES matches (id) ON DELETE CASCADE,
